Spike to try to get MIDI Continuous Controllers working.
Approach isn't ideal - would need a refactor around the state
to combine the 'settings' for controls that can be changed, then
to potentially combine the reducers.
Currently, trying to change the props to be Immutable so that
changes can be calculated, but the controlValues don't appear to
be set by the Midi controllers in the state.
Break debugging is tricky because messages stack up then get dispatch'd
whilst you're in the reducer - throwing an error.