Arpeggiator controlled by the keyboard.
Previous experiment in the synth is not a good idea.
Inspired by http://arpeggiator.desandro.com/
Make one which is triggered by the keyboard.
It uses offsets from the current note so whichever key you
press is the 'root' note.
Allow for rests and 'holds' with notation.
TODO: add proper controls - select arpeggio type, better buttons
to toggle on/off and layout better.
Skips a little when run 'fast' - timing issues or other?