Arpeggiator - fixing layout issues and adding controls.
Fix bug with turning off last note when key up.
Allow changing of pattern & BPM on the fly.
A bit buggy with the number input change - should possibly be
a knob control - typing the number / selecting the input can
break it.
TODO - consider shifting to a control block?
TODO - fix bugginess and allow mouse controls to work with arpeggios.